Move alfa-cli sources to src
