+++ packages/alfa-cli/src/alfa/command/audit/run.ts
@@ -14,12 +14,12 @@ import { Page } from "@siteimprove/alfa-web";
 
 import rules from "@siteimprove/alfa-rules";
 
-import { Profiler } from "../../profiler.js";
+import { Profiler } from "../../profiler";
 
-import type { Arguments } from "./arguments.js";
-import type { Flags } from "./flags.js";
+import type { Arguments } from "./arguments";
+import type { Flags } from "./flags";
 
-import * as scrape from "../scrape/run.js";
+import * as scrape from "../scrape/run";
 
 export const run: Command.Runner<typeof Flags, typeof Arguments> = async ({
   flags,
