+++ packages/alfa-cli/src/alfa/command/scrape/arguments.ts
