+++ packages/alfa-cli/src/alfa/command/scrape/run.ts
@@ -19,8 +19,8 @@ import {
 import { Timeout } from "@siteimprove/alfa-time";
 import { URL } from "@siteimprove/alfa-url";
 
-import type { Arguments } from "./arguments.js";
-import type { Flags } from "./flags.js";
+import type { Arguments } from "./arguments";
+import type { Flags } from "./flags";
 
 export const run: Command.Runner<typeof Flags, typeof Arguments> = async ({
   flags,
