+++ packages/alfa-cli/tsconfig.json
@@ -4,19 +4,7 @@
   "compilerOptions": {
     "resolveJsonModule": true
   },
-  "files": [
-    "package.json",
-    "bin/alfa.ts",
-    "bin/alfa/command/audit.ts",
-    "bin/alfa/command/audit/arguments.ts",
-    "bin/alfa/command/audit/flags.ts",
-    "bin/alfa/command/audit/run.ts",
-    "bin/alfa/command/scrape.ts",
-    "bin/alfa/command/scrape/arguments.ts",
-    "bin/alfa/command/scrape/flags.ts",
-    "bin/alfa/command/scrape/run.ts",
-    "bin/alfa/profiler.ts"
-  ],
+  "files": [],
   "references": [
     { "path": "../alfa-command" },
     { "path": "../alfa-formatter" },
